CloudScripts/python/bulk_cli/requirements.txt
cachetools==4.0.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
httplib2==0.17.0
oauth2client==4.1.3
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
rsa==4.0
six==1.14.0
uritemplate==3.0.1
